+++ seq2science/cli.py
@@ -24,14 +24,15 @@ def _import():
     """
     this function serves that we can do imports as late as possible, for faster auto-completion
     """
-    global webbrowser, contextlib, yaml, psutil, snakemake, logger, setup_logger, xdg
+    global webbrowser, contextlib, yaml, psutil, snakemake, logger, setup_logger, xdg, datetime, _logging
     import yaml
     import psutil
     import webbrowser
     import contextlib
+    import datetime
 
     import snakemake
-    from snakemake.logging import logger, setup_logger
+    from snakemake.logging import logger, setup_logger, _logging
     import xdg
 
 
@@ -309,7 +310,7 @@ def _run(args, base_dir, workflows_dir, config_path):
 
     # run snakemake/seq2science
     #   1. pretty welcome message
-    setup_logger()
+    setup_seq2science_logger()
     log_welcome(logger, parsed_args)
     if not args.skip_rerun or args.unlock:
         #   2. start a dryrun checking which files need to be created, and check if
@@ -543,3 +544,15 @@ def resource_parser(parsed_args):
         # otherwise assume system memory
         mem = psutil.virtual_memory().total / 1024 ** 3
         parsed_args["resources"]["mem_gb"] = round(mem)
+
+
+def setup_seq2science_logger():
+    seq2science_logfile = os.path.abspath(
+        "seq2science."
+        + datetime.datetime.now().isoformat().replace(":", "")
+        + ".log"
+    )
+    setup_logger()
+    logger.logfile = seq2science_logfile
+    logger.logfile_handler = _logging.FileHandler(seq2science_logfile)
+    logger.logger.addHandler(logger.logfile_handler)